Transaction report service

+++ customer/src/main/java/com/example/customer/dto/EmployeeDTO.java
@@ -2,11 +2,13 @@ package com.example.customer.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class EmployeeDTO {
     private String empId;
     private String companyName;
-    private Double ctc;
-    private Double monthlySalary;
+    private BigDecimal ctc;
+    private BigDecimal monthlySalary;
     private Integer experience;
 }
